Changed tests so that they now pass (code relocation + changes made them fail)
